Adjust `ptype` documentation to reflect that only a named list is expected
Previously, we required a data frame ptype and we would cast the unchopped data frame to this ptype. This was confusing, and could add additional columns to the unchopped result that were only present in the ptype. A named list of column specific ptypes is more in line with `pivot_longer()`.