Improve `unchop()` performance (#1140)
* Vectorize `df_unchop()` and have it handle `keep_empty`

Based on the great ideas from tidyverse/tidyr#1126 from @mgirlich. This also allows `unchop()` to handle unchopping df-cols / rcrd-cols (which is a no-op).

* Adjust `ptype` documentation to reflect that only a named list is expected

Previously, we required a data frame ptype and we would cast the unchopped data frame to this ptype. This was confusing, and could add additional columns to the unchopped result that were only present in the ptype. A named list of column specific ptypes is more in line with `pivot_longer()`.

* Apply suggestions from code review

Co-authored-by: Maximilian Girlich <maximilian.girlich@metoda.com>

* Run revdeps

- ffscrapr, r2dii.analysis, strand due to #1142
- simTool, tidyjson due to `unchop()` dropping names. Will fix momentarily.
- Unsure about visR

* Don't zap names when unchopping

This broke a few revdeps, and isn't necessary/correct. `vec_unchop()` combines with `"minimal"` repair by default, so there shouldn't be any issues.

* Re-run revdeps for tidyjson, simTool, and visR

Turns out that visR was also affected by the name zapping behavior

* Drop outer names on the list column

Outer names can be problematic in `vec_unchop()` if the inner elements have size >1, as then some kind of name spec is needed to combine them. This fixes an issue with graphTweets that popped up after removing the `zap()` from `vec_unchop()`, which zapped both outer and inner names.

Co-authored-by: Maximilian Girlich <maximilian.girlich@metoda.com>