Add additional checks to spec columns
In particular, `.name` must be unique. In the case of a `pivot_wider()`, we need to be able to manipulate the column order of the widened columns using the original order supplied in `.name`. This means we can't have name repair altering the names midway through the pivot. For `pivot_longer()`, each `.name` must uniquely identify the row in the spec.