Reorder values columns before combining with id columns (#1236)
* Add additional checks to spec columns

In particular, `.name` must be unique. In the case of a `pivot_wider()`, we need to be able to manipulate the column order of the widened columns using the original order supplied in `.name`. This means we can't have name repair altering the names midway through the pivot. For `pivot_longer()`, each `.name` must uniquely identify the row in the spec.

* Reorder values columns before combining with id columns

Since `.name` was checked to be unique, the values columns shouldn't have any name repair applied to them and should match `.name` exactly. We reorder those to match what the user supplied, then combine with the id columns, at which point we apply name repair right before returning.

* NEWS bullet