Retain row names of data.frames in `df_append()` (#1454)
* Retain row names of data.frames in `df_append()`

This seems reasonable because `reconstruct_tibble()` has no way to restore these, and the number of rows will never change here. `dplyr:::dplyr_col_modify()` does a similar thing. The hlaR package actually relies on this, as they expect row names to persist on a data.frame after a call to `unite()`.

* Add tests