Fix error messages in tests for separate_wider_*() functions'
Fixes #1475. The separate_wider_position() and separate_wider_regex() functions provide the too_few argument to control their behavior when one or more of the strings cannot be separated into as many pieces as the user specifies. When this happens, the error message that is displayed recommends specifying too_few as "start", rather than the valid value "align_start".